simplify implementation, add sortBy indicators
