--- apps/web/app/app.dub.co/(dashboard)/[slug]/(ee)/program/partners/[partnerId]/comments/page-client.tsx
@@ -1,10 +1,0 @@
-"use client";
-
-import { PartnerComments } from "@/ui/partners/partner-comments";
-import { useParams } from "next/navigation";
-
-export function ProgramPartnerCommentsPageClient() {
-  const { partnerId } = useParams() as { partnerId: string };
-
-  return <PartnerComments partnerId={partnerId} />;
-}
