+++ apps/web/app/app.dub.co/(dashboard)/[slug]/(ee)/program/partners/[partnerId]/comments/page.tsx
@@ -1,10 +1,15 @@
-import { ProgramPartnerCommentsPageClient } from "./page-client";
+"use client";
+
+import { PartnerComments } from "@/ui/partners/partner-comments";
+import { useParams } from "next/navigation";
 
 export default function ProgramPartnerCommentsPage() {
+  const { partnerId } = useParams() as { partnerId: string };
+
   return (
     <>
       <h2 className="text-content-emphasis text-lg font-semibold">Comments</h2>
-      <ProgramPartnerCommentsPageClient />
+      <PartnerComments partnerId={partnerId} />
     </>
   );
 }
